agent: renamed new packages and put test packages under testutil
