+++ pkg/agent/rpc/handlers.go
@@ -1,4 +1,4 @@
-package agentservice
+package rpc
 
 import (
 	"fmt"
