+++ pkg/agent/rpc/service.go
@@ -1,4 +1,4 @@
-package agentservice
+package rpc
 
 import (
 	"context"
